+++ CHANGELOG.md
@@ -1,4 +1,4 @@
 > The version number of fuji follows `semver` now: https://semver.org/
 
-- feature: new supported version for Minecraft 1.20.1! (Alpha Version)
+- (command_event module) feature: add `enable` option for each command event. (Thanks to @imfrea)